requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
openai-whisper
soundfile
numpy<2.3
numba
kokoro
bs4
ffmpeg-python
python-multipart
